+++ console/src/components/Common/utils/routesUtils.js
@@ -5,7 +5,7 @@ import {
   checkIfTable,
   getFunctionSchema,
   getFunctionName,
-} from './pgSchemaUtils';
+} from './pgUtils';
 
 /*** DATA ROUTES ***/
